feat(cli): delegate to book_creator orchestrator
