+++ main.py
@@ -1,33 +1,15 @@
 from __future__ import annotations
 
-import os
-import tempfile
+import asyncio
+import sys
 
-from crawler import extract_links
-from logger import get_logger
-from merger import PyPDF2Merger, merge_documents
-from renderer import PlaywrightRenderer, render_to_pdf
-from usecase import create_book
-
-logger = get_logger(__name__)
+from book_creator import run
 
 
-async def run(url: str, output: str, timeout: int = 15000) -> str:
-    """Crawl ``url`` and produce a merged PDF at ``output``."""
-
-    renderer = PlaywrightRenderer()
-
-    async def render_page(u: str, dest: str, t: int) -> bool:
-        return await render_to_pdf(u, dest, timeout=t, renderer=renderer)
-
-    def merge_pdfs(paths: list[str], dest: str) -> bool:
-        return merge_documents(paths, dest, merger=PyPDF2Merger())
-
-    return await create_book(
-        url,
-        output,
-        timeout,
-        link_extractor=extract_links,
-        renderer=render_page,
-        merger=merge_pdfs,
-    )
+if __name__ == "__main__":  # pragma: no cover - manual invocation only
+    if len(sys.argv) < 3:
+        raise SystemExit("Usage: python main.py <url> <output> [timeout]")
+    url = sys.argv[1]
+    output = sys.argv[2]
+    timeout = int(sys.argv[3]) if len(sys.argv) > 3 else 15000
+    asyncio.run(run(url, output, timeout))
