+++ tests/test_main.py
@@ -6,12 +6,12 @@ from unittest.mock import AsyncMock, patch
 ROOT_DIR = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT_DIR))
 
-from main import run  # noqa: E402
+from book_creator import run  # noqa: E402
 
 
-@patch("main.merge_documents")
-@patch("main.render_to_pdf", new_callable=AsyncMock)
-@patch("main.extract_links")
+@patch("book_creator.merge_documents")
+@patch("book_creator.render_to_pdf", new_callable=AsyncMock)
+@patch("book_creator.extract_links")
 def test_run_orchestrates(mock_extract, mock_render, mock_merge, tmp_path):
     mock_extract.return_value.links = ["https://a", "https://b"]
     out = tmp_path / "book.pdf"
